+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Use more-current stripes-components. Refs STRIPES-495.
 * Update dependencies. Refs STRIPES-501. 
 * Provide ID for Scan-ID form for easy access by tests. Refs UITEST-20. Available from v1.1.2. 
+* Add setting for audio alerts in circulation. Fixes UICIRC-47.
 
 ## [1.1.1](https://github.com/folio-org/ui-circulation/tree/v1.1.1) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.1.0...v1.1.1)